requirements.txt
aiofiles==24.1.0
aiogram==3.18.0
aiohappyeyeballs==2.5.0
aiohttp==3.11.13
aiosignal==1.3.2
aiosqlite==0.21.0
annotated-types==0.7.0
appdirs==1.4.4
async-timeout==5.0.1
attrs==25.1.0
beautifulsoup4==4.13.3
certifi==2025.1.31
charset-normalizer==3.4.1
decorating==0.6.1
dotenv==0.9.9
frozenlist==1.5.0
idna==3.10
magic-filter==1.0.12
mal-api==0.5.3
multidict==6.1.0
propcache==0.3.0
pydantic==2.10.6
pydantic_core==2.27.2
python-dotenv==1.0.1
requests==2.32.3
soupsieve==2.6
typing_extensions==4.12.2
urllib3==2.3.0
yarl==1.18.3
